Rename the two "query" commands to "show"
Still provides the original commands by re-invoking knife with the new
way of doing things if it's called the old way.

While we are querying values on the vm and pool, other knife commands
all use "show" instead of "query", so this is an attempt to conform.
