+++ lib/chef/knife/vsphere_vm_query.rb
@@ -1,6 +1,3 @@
-# Author:: Brian Dupras (<bdupras@rallydev.com>)
-# License:: Apache License, Version 2.0
-
 require 'chef/knife'
 require 'chef/knife/base_vsphere_command'
 require 'rbvmomi'
@@ -12,37 +9,10 @@ class Chef::Knife::VsphereVmQuery < Chef::Knife::BaseVsphereCommand
   common_options
 
   def run
-    $stdout.sync = true
-    vmname = @name_args[0]
-    if vmname.nil?
-      show_usage
-      fatal_exit('You must specify a virtual machine name')
-    end
-
-    query_string = @name_args[1]
-    if query_string.nil?
-      show_usage
-      fatal_exit('You must specify a QUERY value (e.g. guest.ipAddress or network[0].name)')
-    end
-
-    vim_connection
-
-    dc = datacenter
-    folder = find_folder(get_config(:folder)) || dc.vmFolder
-
-    vm = traverse_folders_for_vm(folder, vmname) || abort("VM #{vmname} not found")
-
-    # split QUERY by dots, and walk the object model
-    query = query_string.split '.'
-    result = vm
-    query.each do |part|
-      message, index = part.split(/[\[\]]/)
-      unless result.respond_to? message.to_sym
-        fatal_exit("\"#{query_string}\" not recognized.")
-      end
-
-      result = index ? result.send(message)[index.to_i] : result.send(message)
-    end
-    puts result
+    args = ARGV
+    args[2] = 'show'
+    ui.warn 'vsphere vm query is moving to vsphere vm show. Next time, please run'
+    ui.warn args.join " "
+    Chef::Knife.run(args)
   end
 end
